YKI(Frontend): Replace use of tabs in header with navigation links [deploy]
