+++ frontend/packages/yki/public/i18n/sv-SE/common.json
@@ -68,7 +68,8 @@
       "header": {
         "accessibility": {
           "continueToMain": "Fortsätt till innehållet",
-          "langSelectorAriaLabel": "Kieli / Språk / Language"
+          "langSelectorAriaLabel": "Kieli / Språk / Language",
+          "mainNavigation": "Huvudnavigering"
         },
         "sessionState": {
           "activeRegistrations_one": "Du har redan påbörjat 1 anmälan.",
